Don't store the QSGNode pointer
The QSGNode subtree may be cleared at any time. Get the subtree via the
node pointer passed in updatePaintNode, to ensure that the subtree is
still valid each update.

Some references are still being stored but invalidated when a new
subtree is created. QTBUG-34994 has been created to track fixing that.

Task-number: QTBUG-33553
Change-Id: I2115aff931d42b613d207553c636be7d80c405bb
Reviewed-by: Gunnar Sletta <gunnar.sletta@digia.com>