Fix interaction of QQuickItems with the garbage collector
The QObject ownership of QQuickItem objects is not accessible / mutable
in QML, because the parent property maps to the (dynamic) visual parent.

There are use-cases of creating QQuickItem objects without a QObject parent and
to support this, the visual parent needs to mark its visual children in order
to provide intuitive semantics.

[ChangeLog][QtQuick][Import Behavior Changes] A QQuick Item is now strongly
referenced by its visual parent item, so it doesn't require a QObject parent to
stay alive.

Task-number: QTBUG-35913
Change-Id: Ief2d40ac76298a0cf241ca73ff654c4ecfa12748
Reviewed-by: Lars Knoll <lars.knoll@digia.com>