[new compiler] Fix scope, binding target and property object for bindings in group properties
Resolve the confusion about _qobject and _qobjectForBindings, use _qobject for
all property related bits and _qobjectForBindings is gone, in favor of
_scopeObject. Usually they are the same, but in group properties for example,
the _scopeObject remains the outter object but the _qobject becomes the
object that's returned by the group property, i.e.

   anchors.fill: parent

For that binding, the scope object is the outter object but the _qobject
is the value of the anchors property.

Then there's binding target, which is revelant for value type properties, where
it remains the outter object. In theory one could nest a value type property in
a regular group object, in which case _scopeObject would be outter most,
_qobject would be the value type wrapper and _bindingTarget would be the group
object.

Change-Id: Ie3135d951910bf40ece192ab6260d7a9e6e9cdd6
Reviewed-by: Lars Knoll <lars.knoll@digia.com>